services/gateway/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
redis>=5.0.0
python-jose[cryptography]>=3.3.0
pydantic>=2.0
pydantic-settings>=2.0
prometheus-fastapi-instrumentator>=6.0.0
